+++ docs/changes.md
@@ -8,6 +8,7 @@ title: Support Change Log
 ### v2.1.1@dev {#v2-1-1}
 
 * Allow to specify `$prefix` and `$suffix` for `Orchestra\Support\Str::replace()` helper.
+* Implement [PSR-4](https://github.com/php-fig/fig-standards/blob/master/proposed/psr-4-autoloader/psr-4-autoloader.md) autoloading structure.
 
 ### v2.1.0 {#v2-1-0}